edit dtype = float32

+++ tensorflowTUT/tf11_build_network/full_code.py
@@ -22,8 +22,8 @@ def add_layer(inputs, in_size, out_size, activation_function=None):
     return outputs
 
 # Make up some real data
-x_data = np.linspace(-1, 1, 300)[:, np.newaxis]
-noise = np.random.normal(0, 0.05, x_data.shape)
+x_data = np.linspace(-1, 1, 300, dtype=np.float32)[:, np.newaxis]
+noise = np.random.normal(0, 0.05, x_data.shape).astype(np.float32)
 y_data = np.square(x_data) - 0.5 + noise
 
 ##plt.scatter(x_data, y_data)
